Use metrique dep and extract DynamicInflectionEntry
