+++ metrique-util/Cargo.toml
@@ -15,7 +15,7 @@ unexpected_cfgs = { level = "warn", check-cfg = ['cfg(tokio_unstable)'] }
 default = []
 state = ["dep:arc-swap"]
 tokio-metrics-bridge = [
-    "dep:metrique-writer-core",
+    "dep:metrique",
     "dep:tokio",
     "dep:tokio-metrics",
     "dep:tracing",
@@ -23,11 +23,10 @@ tokio-metrics-bridge = [
 
 [dependencies]
 metrique-core = { path = "../metrique-core", version = "0.1.18" }
+metrique = { path = "../metrique", version = "0.1.23", optional = true, default-features = false }
 arc-swap = { version = "1", optional = true }
-metrique-writer-core = { path = "../metrique-writer-core", version = "0.1.13", optional = true }
 tokio = { workspace = true, optional = true, features = ["time", "rt"] }
-# TODO: Update before merge, should point to tokio-metrics on crates.io
-tokio-metrics = { git = "https://github.com/yulnr/tokio-metrics", branch = "metrique-entry-support", optional = true, features = ["rt", "metrique-integration"] }
+tokio-metrics = { version = "0.5.0", optional = true, features = ["rt", "metrique-integration"] }
 tracing = { workspace = true, optional = true }
 
 [dev-dependencies]
