+++ metrique-util/src/lib.rs
@@ -11,6 +11,8 @@ mod state;
 pub use state::{LatestRef, State};
 
 #[cfg(feature = "tokio-metrics-bridge")]
+mod dynamic_inflection;
+#[cfg(feature = "tokio-metrics-bridge")]
 mod tokio_metrics_reporter;
 #[cfg(feature = "tokio-metrics-bridge")]
 pub use tokio_metrics_reporter::{
